+++ CHANGES.md
@@ -45,6 +45,7 @@ Changes in 11.0
 - SQL calculate Class Rank for Progress Periods in ReportCards.php
 - SQL Fix replacement in case the "username" contains the prefix in Removeaccess.php, thanks to @0xD0M1M0
 - SQL add accounting_categories table in rosariosis.sql & rosariosis_mysql.sql
+- Add Categories program to Accounting module in Categories.php, Menu.php, rosariosis.sql & rosariosis_mysql.sql
 
 Changes in 10.9.7
 -----------------